+++ lib/scanner/core.rb
@@ -14,10 +14,15 @@ module Yawast
 
           print_header
 
+          if options.output != nil
+            Yawast::Shared::Output.setup @uri, options
+          end
+
           ssl_redirect = Yawast::Scanner::Plugins::SSL::SSL.check_for_ssl_redirect @uri
           if ssl_redirect
             @uri = ssl_redirect
             puts "Server redirects to TLS: Scanning: #{@uri}"
+            Yawast::Shared::Output.log_value 'server_tls_redirect', @uri
           end
 
           Yawast::Scanner::Plugins::SSL::SSL.set_openssl_options
@@ -42,6 +47,7 @@ module Yawast
 
           #cache the HEAD result, so that we can minimize hits
           head = get_head
+          Yawast::Shared::Output.log_hash 'http', 'head', 'raw', head.to_hash
           Yawast::Scanner::Generic.head_info(head, @uri)
 
           #perfom SSL checks
@@ -75,6 +81,7 @@ module Yawast
           # less than 24 hours. if a scan is that long, we have bigger problems
           elapsed_time = Time.at(Time.now - start_time).utc.strftime('%H:%M:%S')
 
+          Yawast::Shared::Output.write_file
           puts "Scan complete (#{elapsed_time})."
         rescue => e
           Yawast::Utilities.puts_error "Fatal Error: Can not continue. (#{e.class}: #{e.message})"