+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.7.0 - In Development
 
+* [#134](https://github.com/adamcaudill/yawast/issues/134) - Add options to DNS command
+
 ## 0.6.0 - 2018-01-16
 
 * [#54](https://github.com/adamcaudill/yawast/issues/54) - Check for Python version in Server header